save the networks and the models
I save the networks and the model into two folders, networks and outputs.

Into the folder we put two file of parameters. In networks folder the parameters file describe the interval between two saving and which networks we want save.

In outputs parameters file we save only the interval

I changed the interface: I add a button two reload a model (.world file) and a report for actual tick

It's not a good idea to load only a network, because is really complex to set the model. So, in this branch II put only the "load-model" feature

I use nw::setcontext with turtles and the name of the link agentset. It's necessary to use turtles because some links come from a breed to another breed

If the interval is set to zero means that the saving not occur

We can view the networks (graphml  fornmat) with https://jsonformatter.org/xml-viewer