Optimization
First attempt at optimization
Instead of calculating  c at each step (with file access, ahem, that must go as well) we precook it. But we eat it only if it is fresh.

However, this broke my new test, so this also is to manage some issue with it and to test the [ci skip] directive.

[ci skip]