first commit simone
added two monitors in the interface to control how many commited a crime and how many links we have between them. A new link breed is added: the goal is to merge family-friendship-professional links into a weighted connection.
Agents also have a size property that follows a new variable called crime-activity, which goes up to 5 when a crime is committed and then goes back to 1 after 4 ticks, so that we can trace the criminal activity of agents.