Introduced oc-status
We get rid of the n^2 / (n+1)^2 calculation.

Instead, we calculate now a score based on the number of links of all kind to OC neighbors, plus the number of criminal acts done together. We subratct the minimum (thus starting from zero) and we set the mean to 1. The max will depend on the range.

This number is then used to multiply the arrest probability.

This will produce the correct number of arrests given that all members participate equally to crimes.